+++ src/ui/StatusBar.java
@@ -58,9 +58,9 @@ public class StatusBar {
 		//Draw text
 		
 		//Time Count
-		int time = GameManager.getInstance().getTimeCount();
-		int sec = time%60;
-		int min = time/60;
+		int time = (int) GameManager.getInstance().getTimeCount();
+		int sec = time % 60;
+		int min = time / 60;
 		gc.setFill(Color.BLACK);
 		gc.setFont(TIME_FONT);
 		gc.setTextAlign(TextAlignment.CENTER);
